services/post/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
grpcio
grpcio-tools
protobuf
sqlalchemy
